Rebase onto develop and fix conflicts

+++ Market/Pages/SalePage.xaml.cs
@@ -36,7 +36,7 @@ namespace Market
                 {
                     var temp = queryPrd.Find(i);
                     var tempp = queryStock.Find(temp.Barcode);
-                    int amountt = tempp.Amount;
+                    int amountt = (int)tempp.Amount;
                     if(amountt > 100) { amountt = 100; }
                     items.Add(new StockItem() { Title = temp.Name, Completion = amountt });
                 }
